refactored independent filesystem code into Filesystem.{h,cpp}
