+++ Utilities/Loader.h
@@ -85,5 +85,3 @@ private:
 
   void printOrganism(long); // strictly to debug all_organisms entries
 };
-
-bool isDirectory(const std::string&);
